fix: delete student and course

+++ app-back/src/services/studentService.ts
@@ -1,6 +1,7 @@
 import { TypeStudent } from '../types/StudentTypes';
 import * as studentRepository from '../repositories/studentRepository';
-import { conflictError, notFoundError, unprocessableEntity } from '../utils/errorUtils';
+import * as enrollmentRepository from '../repositories/enrollmentRepository';
+import { conflictError, notFoundError, unauthorizedError, unprocessableEntity } from '../utils/errorUtils';
 
 export async function getStudents() {
     return await studentRepository.getStudents();
@@ -22,5 +23,7 @@ export async function updateStudent(id: number, data: TypeStudent) {
 export async function deleteStudent(id: number) {
     const studentExists = await studentRepository.findById(id);
     if (!studentExists) throw notFoundError("Student not exists!");
+    const enrolledStudent = await enrollmentRepository.getEnrollmentByStudentId(id);
+    if (enrolledStudent.length > 0) throw unauthorizedError("The student is enrolled in a course.")
     await studentRepository.deleteStudent(id);
 }
